lib: utils/regmap: Use FDT node offset as regmap ID
Since the FDT is not modified during driver initialization, node offsets
are just as suitable as phandles for use as identifiers: they are stable
and unique. With this change, it is no longer necessary to pass the
phandle to the driver init functions, so these init functions now use
the same prototype as other kinds of drivers.

This matches what is already done for I2C adapters.

Series-to: opensbi@lists.infradead.org

Cover-letter:
Unify driver cold init function prototypes
Currently two kinds of FDT drivers (GPIO and regmap) use a different
prototype for their driver init functions, additionally passing in the
phandle of the FDT node. However, it turns out that using the phandle is
not necessary. It's only used as a unique ID for a device instance, but
the FDT node offset works just as well for this purpuse, and in fact the
I2C drivers use the FDT node offset as the device ID already.

By removing the phandle argument from these driver init functions, all
FDT drivers now use the same init function prototype, which will allow
further code reuse in the future.
END

Signed-off-by: Samuel Holland <samuel.holland@sifive.com>