platform: Drop IPI warm init and exit hooks
Now that the SBI IPI core clears IPIs at warm boot in a generic way,
none of the drivers or platforms use these hooks, and we can remove
them. Platforms need only to initialize the driver once during cold
init. If other hooks are needed in the future, they can be added to
struct sbi_ipi_device.

Series-to: opensbi@lists.infradead.org

Series-version: 2

Cover-letter:
Manage IPI driver lifecycle from SBI core
For three kinds of devices (IPI, irqchip, and timer) the platform is
responsible for implementing extra warm init and exit lifecycle hooks.
Since the driver lifecycle is the same regardless of platform, I think
it makes sense to call these lifecycle functions from the SBI core, not
the platform. This produces smaller code since we don't need to track
the current driver in multiple places. And since it makes the FDT driver
interface the same for all kinds of devices, it allows us to reuse the
same FDT scan/match/init code across all of them.

This series converts the IPI drivers. This is the most trivial of the
three, since all of the logic in the existing IPI .warm_init hooks is
actually a duplicate of the .ipi_clear hook in struct sbi_ipi_device.

I sent the timer conversion here[1], and the irqchip conversion is
upcoming and is the largest of the three.

[1]: http://lists.infradead.org/pipermail/opensbi/2024-September/007332.html
END

Signed-off-by: Samuel Holland <samuel.holland@sifive.com>